requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
datasets==1.9.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
dj-database-url==0.5.0
Django==3.2.5
entrypoints==0.3
filelock==3.0.12
fsspec==2021.6.1
gunicorn==20.1.0
huggingface==0.0.1
huggingface-hub==0.0.12
idna==2.10
img2pdf==0.4.1
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
multiprocess==0.70.12.2
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.2
pdf2image==1.16.0
pdfminer.six==20201018
pickleshare==0.7.5
pikepdf==3.0.0
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
pyarrow==4.0.1
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.9.0
PyMuPDF==1.18.14
pyparsing==2.4.7
pyrsistent==0.18.0
pytesseract==0.3.8
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.7.6
requests==2.25.1
sacremoses==0.0.45
Send2Trash==1.7.1
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
tokenizers==0.10.3
torch==1.9.0+cpu
torchaudio==0.9.0
torchvision==0.10.0+cpu
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
transformers==4.8.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
wincertstore==0.2
xxhash==2.0.2
